==> 04_model-analysis_load-infection-endpoint.Rmd <==
---
title: "04_model-analysis_infection-endpoint_load"
author: "Bryony Allen"
date: "14/05/2019"
output: 
  word_document:
    toc: true 
    toc_depth: 2 
---

NOTE THIS SCRIPT FILE HAS BEEN REDUCED DOWN TO ESSENTIALS > SEE "full-process" FILE IN "code" FOLDER FOR ALL ANALYSIS CHECKS 


```{r packages, include=FALSE, warning=FALSE, error=FALSE, message=FALSE}
# Install & load in required packages 

# for tidying data  
library(dplyr)
# library(tidyr)
# library(tibble)

# for plotting 
library(ggplot2) 
library(lattice)
library(ggridges)

# for stats 
library(binom)
library(multcomp)  ### this masks dplyr::select !!!! 

# library(purr)  # purr guidance > for evaluating models http://ijlyttle.github.io/isugg_purrr/presentation.html#(1) 
# library(broom)   # broom summarizes key information about models in tidy tibble()s >> https://github.com/tidymodels/broom
```

# Part 1: Bd: import data 

> **N.B.**   At the moment this code chuck reads in a .csv file with endpoint infection data, originally compiled in excel and then cleaned in "02_tidy_data-qpcr"script. In the future you will read in the .csv file of merged qPCR outputs (created, checked and cleaned in "02_tidy_data-qpcr"script) and experiment metadata (checked and cleaned in "02_tidy_data-metadata"script). 

<br>
```{r import data, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE}
data.endpoint <- read.csv('data/02_clean-data.csv') 
    data.endpoint$ExperimentNo <-  as.factor(data.endpoint$ExperimentNo)   # makes experiment number a factor 

glimpse(data.endpoint)   #check how the dataset has imported (aka whether it has the data type right)  
```

# Part 2: Bd: visualise endpoint infection load data 
```{r graph labels, include=FALSE, warning=FALSE, error=FALSE, message=FALSE}
# make label vectors to use in plots
sp.labs <- c(Bb = "Bufo bufo", Rt = "Rana temporaria", Am = "Alytes muletensis")
exp.labs <- c("1" = "Bufo bufo I", "2" = "Bufo bufo II", "3" = "Rana temporaria", "4" = "Alytes muletensis","5" = "Alytes muletensis II")
expos.labs <- c("1" = "Bd", "2" = "Rv", "3" = "Bd-Rv", "4"="Rv-Bd")
```

Here I only include Bd infection loads over the 0.1 GE threshold. 

```{r Bd Bd boxplot, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.cap= "Fig.1. Boxplot of the endpoint Bd load for each treatment group across the three host species. Bd load is quantified as genomic equivalents (GE) where 1 GE represents 1 Bd zoospore. Where the black dots represent each sample and the red diamond the mean for that group.", fig.width=16, fig.height=8}

plot.Bd.GE.sp <- data.endpoint %>% 
   filter((Bd.endpoint.status=='1' & Bd.endpoint.GE > 0.1))  %>%   # filter so its just Bd infection that are over the threshold 
      ggplot(aes(interaction(x= Treatment, ExperimentNo), y= Bd.endpoint.GE, fill = ExperimentNo)) +
        geom_boxplot() +
        stat_summary(fun.y="mean", geom="point", shape=23, size=3, fill="red") +    # adds a marker for mean - for normally distibuted data the median and mean will match closely 
        geom_jitter(width = 0.05) +   # adds the data points  (use colour to change this )
        scale_y_log10() +  # logs the axis 
        scale_x_discrete(labels=c("Bd", "Bd>Rv", "Rv>Bd","Bd", "Bd>Rv", "Rv>Bd","Bd", "Bd>Rv", "Rv>Bd", "Bd", "Bd>Rv", "Rv>Bd", "Bd")) +
        xlab("Treatment Group") +
        ylab("Bd load\n(GE)") +
        theme(axis.title.y = element_text(angle=0, size=14), legend.text = element_text(face = "italic")) + 
        scale_fill_grey(start=1, end=.5, labels=exp.labs) +
        theme_bw()
        
 plot.Bd.GE.sp + theme(legend.key=element_blank()) +theme(legend.text = element_text(face = "italic")) +theme(axis.text.x = element_text(angle=30, hjust=1,vjust=1)) + theme(axis.title = element_text(size=16))
```


<br>

```{r distribution Bd GE, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.cap= "Fig. 2a. The distribution of Bd load expressed as GE from endpoint tissue samples.", fig.width=16, fig.height=6}

data.endpoint %>%     # ridge plot to see distribution of data  
    filter((Bd.endpoint.status=='1' & Bd.endpoint.GE > 0.1))  %>%  ggplot(aes(Bd.endpoint.GE, Treatment)) +
      geom_density_ridges(
                      jittered_points = TRUE, quantile_lines = TRUE, scale = 0.9, alpha = 0.7,
                      vline_size = 1, vline_color = "red",
                      point_size = 0.4, point_alpha = 1,
                      position = position_raincloud(adjust_vlines = TRUE)
          ) + theme_ridges()
```

<br>

... and when Bd load is logged 

```{r log distribution Bd GE, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.cap= "Fig. 2b. The logged distribution of Bd load expressed as genomic equivalents (GE), where one GE is equivalent to a single zoospore, from endpoint tissue samples.", fig.width=16, fig.height=6}
data.endpoint %>%     # ridge plot to see distribution of data  
    filter((Bd.endpoint.status=='1' & Bd.endpoint.GE > 0.1))  %>%  ggplot(aes(log(Bd.endpoint.GE), Treatment)) +
      geom_density_ridges(
                      jittered_points = TRUE, quantile_lines = TRUE, scale = 0.9, alpha = 0.7,
                      vline_size = 1, vline_color = "red",
                      point_size = 0.4, point_alpha = 1,
                      position = position_raincloud(adjust_vlines = TRUE)
          ) + theme_ridges()
```


<br>

```{r Bd summary}

Bd.infect.sum <- data.endpoint %>%
  filter((Bd.endpoint.status=='1' & Bd.endpoint.GE > 0.1))  %>% 
  group_by(ExperimentNo, Treatment) %>%  
    summarize(avg=mean(Bd.endpoint.GE), n=n(), sd=sd(Bd.endpoint.GE), se=sd/sqrt(n))


Rv.infect.sum <- data.endpoint %>%
  group_by(ExperimentNo, Treatment) %>%  
    summarize(avg=mean(var2), n=n(), sd=sd(var2), se=sd/sqrt(n))
```



# Part 3: Bd: apply models to endpoint infection load data 

GLM's where    
<br>
    response variable = endpoint load [continuous; ]   
  <br>
    explanatory variable(s) = Treatment [categorical, levels = 4] & Species [categorical, levels = 5]

<br>
  
  
**N.B.**  I use ExperimentNo as a proxy for species where

Experiment No. | Species              | total Bd zsp's   | min. temp. (^o^C) | max. temp. (^o^C)
---------- |------------------------- | -------------| -------------|------------- 
1 | *Bufo bufo* I          | 3,675,000    | 16.6        | 23.5     
2 | *Bufo bufo* II         | 1,443,750   | 16.7        | 27.6    
3 | *Rana temporaria*     | 2,336,250     | 16.7         | 27.6    
4 | *Alytes muletensis* I  | 472,500      | 15     | 16.6 
5 | *Alytes muletensis* II | 294,759      | 15     | 16.6 

... as this also accounts for Bd dose and room temperature variation between experiments. 


Here I create a dataframe with only individuals that am infected with Bd, at a level over the detection threshold of 0.1GE. Note that *Alytes muletensis* II have been removed as they only have one treatment group (Bd only). 

<br>
```{r Bd load df, results='hide', warning=FALSE, error=FALSE, message=FALSE}
Bd.load <- data.endpoint %>%
  filter(!ExperimentNo=='5')  %>%   # removal of Alytes babies
  filter((Bd.endpoint.status=='1' & Bd.endpoint.GE > 0.1))  %>% 
  dplyr::select(ID, Species, ExperimentNo, Scenario, Treatment, Bd.endpoint.status, Bd.endpoint.GE) 

droplevels(Bd.load)
```

## Part 3a: Bd: model checks 

Here I check the chosen model ``` Bd.load4 <- glm(Bd.endpoint.GE ~ ExperimentNo, data=Bd.load, family="Gamma"(link='log'))``` 
 
```{r Bd - glm - summary, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE}
Bd.load.glm <- glm(Bd.endpoint.GE ~ ExperimentNo, data=Bd.load, family="Gamma"(link='log'))

summary(Bd.load.glm)   # Residual deviance: 340.74  on 192 df   AIC: 679

Bd.load.coef <-summary.glm(Bd.load.glm)$coefficients
```

```{r Bd - glm - diganostic plot, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE}
par(mfrow=c(2,2), mar=c(3,3,3,1), mgp=c(2,0.8,0))
plot(Bd.load.glm)
```

## Part 3b: Bd: model fit 

```{r Bd fitted values, echo=FALSE, fig.cap= "Fig. 6. Plot of relationship between the fitted values from the model and the true data values."}

Bd.load.glm <- glm(Bd.endpoint.GE ~ ExperimentNo, data=Bd.load, family="Gamma"(link='log'))

# To see the fitted values from a regression object (the values of the dependent variable predicted by the model), access the ```fitted.values``` attribute from a regression object with ````$fitted.values```  Look at https://bookdown.org/ndphillips/YaRrr/linear-regression-with-lm.html

Bd.load$glm.fit <- Bd.load.glm$fitted.values  # add logisitic fitted values back to the dataframe as a new col

plot(x = Bd.load$Bd.endpoint.GE, y = Bd.load$glm.fit,  # 
     xlab = "True Values",
     ylab = "Model Fitted Values",
     main = "Regression fits of Bd load")

abline(b = 1, a = 0)   # values should fall around this line!

```


> **To Do** Figure this one out. Values should fall around the line but they seem to be clumped by ExperimentNo. tried to do this with the null model but that just grouped everything 

## Part 3d: Bd: model plotting   

To plot the model you need a range of values for which to produce fitted values. Then use the ```predict()``` function to create the model for all the values. ```predict()``` gives you the predicted values based on your (fitted) linear model, the argument `type="response"` will give you the predicted probabilities on the scale of response variable. It back transforms the prediction out of the link scale (inverses the link),so the prediction is on same scale as the dependent variable. 

```{r Bd - glm - predicted values, include=TRUE, warning=FALSE, error=FALSE, message=FALSE}

# create a dataframe of "new" data 
newdat <- expand.grid(ExperimentNo=c("1", "2", "3", "4"),Treatment=c("Bd", "Bd-Rv", "Rv-Bd"))

# predict the value/result of the new data using the glm
newdat <-cbind(newdat, predict(object = Bd.load.glm,   # the model 
                               newdata=newdat, se=TRUE, type="response", print.matrix=T))  # dataframe of new data 
newdat

expl.var <- c(1:3) # chose the range for the x-axis (Experiment No.)
exp.labs <- c("1" = "Bufo bufo I", "2" = "Bufo bufo II", "3" = "Rana temporaria", "4" = "Alytes muletensis I")

newdat1<- subset(newdat, ExperimentNo== "1")    # need to subset the data so you can plot each seperatly 
newdat2<- subset(newdat, ExperimentNo=="2")
newdat3<- subset(newdat, ExperimentNo=="3")
newdat4<- subset(newdat, ExperimentNo=="4")
newdat5<- subset(newdat, ExperimentNo=="5")
```


```{r Bd - glm - plot, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.cap= "Fig. 7. Probability of Bd infection load (GE) predicted by model"}

Bd.load.predict <- ggplot(newdat, aes(x= expl.var, y= fit, color=ExperimentNo)) +       # plot model estimates, color= the data you subsetted by
  geom_line(data = newdat1, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat1
  geom_errorbar(data = newdat1, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1)  +      # error bars for subset newdat1
  geom_line(data = newdat2, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat2
  geom_errorbar(data = newdat2, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1)  +      # error bars for subset newdat2
  geom_line(data = newdat3, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat3
  geom_errorbar(data = newdat3, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1)  +      # error bars for subset newdat3
  geom_line(data = newdat4, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat4
  geom_errorbar(data = newdat4, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1) +       # error bars for subset newdat4
    scale_x_continuous(breaks=seq(1:3),labels=c("Bd", "Bd-Rv", "Rv-Bd"))   # sets the breaks at 1,2 and 3 which correspond to the label names

Bd.load.predict <- Bd.load.predict + 
  labs(title = "glm(Bd.endpoint.load (GE) ~ ExperimentNo, \n family= Gamma(link='log'))", x = "Treatment Group", y = "Bd load predictions\n(fit)", color = "Species\n") +
 scale_color_hue(labels = c("Bufo bufo I", "Bufo bufo II", "Rana temporaria", "Alytes muletensis I")) +
  theme_bw() +
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black")) 

Bd.load.predict

#  ggsave("figs/04_Bd_load-predict.png", plot= Bd.load.predict, device=NULL)
 
```

The Bd load predictions are interesting... 

The two *Bufo bufo* experiments have the highest predicted loads, and reflect the Bd dosed (1 = 3,675,000 zsp's; 2 = 1,443,750 zsp's). As anticipated from the Bd status data and predictions on species susceptibility the *Rana temporaria* loads are low. 

**To Do**  check the model summary to guage difference between the *Bufo bufo* experiments.  

Oddly, the *Alytes muletensis* load prediction is lower than I had expected based off of the Bd status predictions. Similar numbers of *Alytes* and *Bufo* are infected with Bd but the loads differ. Two possibilities here, species susceptibility differences in pathogen burden or (more likely) a reflection of Bd dose (472,500 zsp's compared to millions of zsp's). 


A reminder, just in case ... 

Experiment No. | Species              | total Bd zsp's   | min. temp. (^o^C) | max. temp. (^o^C)
---------- |------------------------- | -------------| -------------|------------- 
1 | *Bufo bufo* I          | 3,675,000    | 16.6        | 23.5     
2 | *Bufo bufo* II         | 1,443,750   | 16.7        | 27.6    
3 | *Rana temporaria*     | 2,336,250     | 16.7         | 27.6    
4 | *Alytes muletensis* I  | 472,500      | 15     | 16.6 
5 | *Alytes muletensis* II | 294,759      | 15     | 16.6 

<br>
**QUESTION** all Bd doses are consisted high but are we seeing dose-dependent Bd loads within this high dose class? 

<br>

# Part 4: Bd: post-hoc tests 

Tukey's honest significant difference 

```{r Bd - posthoc , echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.width=8, fig.height=5}
# Bd.load4 <- glm(Bd.endpoint.GE ~ ExperimentNo, data=Bd.load, family="Gamma"(link='log'))

Tukey.Bd.load <- glht(Bd.load.glm, linfct=mcp(ExperimentNo='Tukey'))
summary(Tukey.Bd.load) 
```
<br>
```{r Bd - posthoc - plot , echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.width=8, fig.height=5}

par(mar=c(2,8,2,1))   #change the size of the margins the first number referes to the bottom margin, the second to the left margin, and so on in a clockwise fashion
plot(Tukey.Bd.load)

# export plot as .png 
#png("figs/04_Bd_load-posthoc.png", type="cairo", units="in", width= 16, height=8, pointsize= 12, res=1000)
#print(plot(Tukey.Bd.load))
#dev.off()
```

<br>

Where 
"1" = "Bufo bufo I",
"2" = "Bufo bufo II", 
"3" = "Rana temporaria", 
"4" = "Alytes muletensis"

**Conclusions**
The Tukey summary suggests everything is significant apparent from the two *Bufo* experiments. 

The plots suggest: 

 * 2-1 : no significant difference between *Bufo bufo* experiments 
 * 3-2 & 3-1 : significant difference between *Rana* and both *Bufo bufo* experiments (to an equal extent)
 * 4-2 & 4-1: significant difference between *Alytes* and both *Bufo bufo* experiments (to an equal extent) <<< see above for my thoughts on this 
 * 4-3: significant (+ve) difference between *Alytes* and *Rana* 


# Part 4: Bd: raw numbers 

```{r Bd numbers}

 <- data.endpoint %>%

  group_by(cit) %>%
summarize(avg=mean(XXXX), n=n(), sd=sd(var2), se=sd/sqrt(n))


var1<-sample(c('red', 'green'), size=10, replace=T)
var2<-rnorm(10, mean=5, sd=1)
df<-data.frame(var1, var2)
df %>% 
group_by(var1) %>% 
summarize(avg=mean(var2), n=n(), sd=sd(var2), se=sd/sqrt(n))

```




# Part 5: Rv: visualise endpoint infection load data 

For the plots below I only plot Rv infection loads over 0.1 and filter out one extremely high Rv viral load (greater than 2000). At the moment I am using the threshold of 0.1 as there are a few infections with 0.000xxx that are skewing the data and unreliable 

**To Do**  At the moment there are 3 samples missing EBF3N qPCR data so as a temporary fix I have averaged the EBF3N score for all the samples and used that to calculate a rough viral load for these 3 samples

> **N.B.** previous thoughts on filtering out samples with negative infection status vs. including and using log(n+1). AF response "usually settle for running one binomial ‘status’ (presence/absence of infection) analysis, and a separate ‘intensity’ (load among infecteds only) analysis.  Zero-inflated/Hurdle models effectively do this in a single model, but I typically think splitting into 2 separate analyses is clearer and easier to explain/interpret."

```{r boxplot Rv viral load, echo=FALSE, fig.cap= "Fig.3. Boxplot of the endpoint Ranaviral load (excluding uninfected individuals) for each treatment group across the three host species. Viral load has been normalised using Leung et al.'s (2017) method. The black dots represent each sample.", fig.width=16, fig.height=8}

plot.Rv.load.sp <- data.endpoint %>% 
 filter(Rv.endpoint.load > 0.1 & !Rv.endpoint.load > 2000) %>%  # filtering out the extremely high Rv viral load so that we can see the data better 
  ggplot(aes(x = interaction(Treatment, ExperimentNo), y= Rv.endpoint.load, fill = ExperimentNo)) + 
        geom_boxplot() +
        #stat_summary(fun.y="mean", geom="point", shape=23, size=3, fill="red") +    
        geom_jitter(width = 0.05) +   # adds the data points       
        xlab("Treatment Group") +
        ylab("Rv viral load\n (log)") +
        theme(legend.text = element_text(face = "italic")) + 
        scale_fill_grey(start=1, end=.5, labels=exp.labs) +
        scale_x_discrete(labels=c("Bd>Rv", "Rv>Bd", "Bd>Rv", "Rv","Rv>Bd", "Bd>Rv", "Rv>Bd","Bd>Rv","Rv","Rv>Bd")) +
        labs(fill="Species")+
        theme_bw()

plot.Rv.load.sp +  theme(legend.key=element_blank()) +theme(legend.text = element_text(face = "italic")) +theme(axis.text.x = element_text(angle=30, hjust=1,vjust=1)) + theme(axis.title = element_text(size=16))

```

The distribution of Rv viral load (excluding uninfected individuals)

```{r distribution Rv load, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.cap= "Fig. 4a. The  distribution of Rv load from endpoint tissue samples.", fig.width=16, fig.height=6}
data.endpoint %>%     # ridge plot to see distribution of data  
    filter(Rv.endpoint.load > 0.1 & !Rv.endpoint.load > 2000)  %>%  ggplot(aes(Rv.endpoint.load, Treatment)) +
      geom_density_ridges(
                      jittered_points = TRUE, quantile_lines = TRUE, scale = 0.9, alpha = 0.7,
                      vline_size = 1, vline_color = "red",
                      point_size = 0.4, point_alpha = 1,
                      position = position_raincloud(adjust_vlines = TRUE)
          ) + theme_ridges()
```

...and when Rv viral load is logged because the distribution of Rv load is skewed. Again excluding uninfected individuals.

```{r log distribution Rv load, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.cap= "Fig. 4b. The logged distribution of Rv load from endpoint tissue samples.", fig.width=16, fig.height=6}
data.endpoint %>%     # ridge plot to see distribution of data  
    filter(Rv.endpoint.load > 0.1 & !Rv.endpoint.load > 2000)  %>%  ggplot(aes(log(Rv.endpoint.load), Treatment)) +
      geom_density_ridges(
                      jittered_points = TRUE, quantile_lines = TRUE, scale = 0.9, alpha = 0.7,
                      vline_size = 1, vline_color = "red",
                      point_size = 0.4, point_alpha = 1,
                      position = position_raincloud(adjust_vlines = TRUE)
          ) + theme_ridges()
```


Just eyeballing this the Rv-Bd group has the highest infection loads compared to the other treatment groups.
?? *QUESTION**  is the sequence/timing of the Rv exposure allowing enough time for Rv loads to build or is it the nauture of that coinfection combination?? 


# Part 6: Rv: apply models to endpoint infection load data 

Here I create a dataframe with only individuals that am infected with Rv 

**To Do** check whether there is a threshold for ranavirus detection??? 

<br>
```{r Rv load df, results='hide', warning=FALSE, error=FALSE, message=FALSE}

Rv.load <- data.endpoint %>%
  filter((Rv.endpoint.load > 0))  %>% 
  dplyr::select(ID, Species, ExperimentNo, Scenario, Treatment, Rv.MCPendpoint.status, Rv.EBF3Nendpoint.status, Rv.endpoint.load) 

droplevels(Rv.load)
```

## Part 6b: Rv: model checks 

Here I check the model with the best fit  ```glm(Rv.endpoint.load ~ Treatment, data=Rv.load, family="Gamma"(link='log'))```
Gamma with log-link has the best (QQ) plots and lowest residual deviance 

```{r Rv - glm - summary, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE}

Rv.load.glm <- glm(Rv.endpoint.load ~ Treatment, data=Rv.load, family="Gamma"(link='log'))
summary(Rv.load.glm)   # Residual deviance: 24.287  on 16 df   AIC: 253.89

Rv.load.coef <-summary.glm(Rv.load.glm)$coefficients
```

```{r Rv - glm - diagnostic plot, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE}
par(mfrow=c(2,2), mar=c(3,3,3,1), mgp=c(2,0.8,0))
plot(Rv.load.glm)
```

```{r Rv fitted values, echo=FALSE, fig.cap= "Fig. 6. Plot of relationship between the fitted values from the model and the true data values."}

Rv.load$glm.fit <- Rv.load.glm$fitted.values  # add logisitic fitted values back to the dataframe as a new col
```

**To Do** split Rv load data into categories low-high or similar? 

## Part 6d: Rv: model plotting

```{r Rv - glm - predicted values, include=TRUE, warning=FALSE, error=FALSE, message=FALSE}
# create a dataframe of "new" data 
newdat <- expand.grid(ExperimentNo=c("1", "2", "3", "4", "5"),Treatment=c( "Bd-Rv", "Rv-Bd"))

# predict the value/result of the new data using the glm
Rv.newdat <-cbind(newdat, predict(object = Rv.load.glm,   # the model 
                               newdata=newdat, se=TRUE, type="response", print.matrix=T))  # dataframe of new data 
Rv.newdat

expl.var <- c(1:2) # chose the range for the x-axis (Treatment)
exp.labs <- c("1" = "Bufo bufo I", "2" = "Bufo bufo II", "3" = "Rana temporaria", "4" = "Alytes muletensis I")

newdat1<- subset(Rv.newdat, ExperimentNo== "1")    # need to subset the data so you can plot each seperatly 
newdat2<- subset(Rv.newdat, ExperimentNo=="2")
newdat3<- subset(Rv.newdat, ExperimentNo=="3")
newdat4<- subset(Rv.newdat, ExperimentNo=="4")
```


```{r Rv - glm - plot, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.cap= "Fig. 7. Probability of Rv infection load predicted by model"}

Rv.load.predict <- ggplot(Rv.newdat, aes(x= expl.var, y= fit)) +       
    geom_line(data = newdat1, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat1
  geom_errorbar(data = newdat1, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1)  +      # error bars for subset newdat1
  geom_line(data = newdat2, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat2
  geom_errorbar(data = newdat2, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1)  +      # error bars for subset newdat2
  geom_line(data = newdat3, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat3
  geom_errorbar(data = newdat3, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1)  +      # error bars for subset newdat3
  geom_line(data = newdat4, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat4
  geom_errorbar(data = newdat4, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1) +       # error bars for subset newdat4
    scale_x_continuous(breaks=seq(1:2),labels=c("Bd-Rv","Rv-Bd")) +   # sets the breaks at 1,2 and 3 which correspond to the label names
  ggtitle("glm(Rv.endpoint.load ~ Treatment,\n family=Gamma(link='log'))") 
      
Rv.load.predict <- Rv.load.predict + 
  ylab("Rv load predictions\n(fit)") +            
  xlab("Treatment Group") +
  theme_bw() +
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black")) 

Rv.load.predict

#  ggsave("figs/04_Rv_load-predict.png", plot= Rv.load.predict, device=NULL)
 
```


?? **QUESTION** it is not incorporating the Rv only group. There are three samples with infection in the Rv only groups but they all have unreliable qPCR history   

<br>

# Part 7: Rv: post-hoc tests 

Tukey's honest significant difference 

```{r Rv - posthoc - , echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.width=8, fig.height=5}
Tukey.Rv.load <- glht(Rv.load.glm, linfct=mcp(Treatment='Tukey'))
summary(Tukey.Rv.load) 
```
<br>

**Conclusions**
The Tukey summary suggests there is significant difference between the two coinfection treatment groups. The plot was pointless. 

?? **QUESTION** I have seen a paper suggest that for low sample sizes a Welch’s modified t-tests is better????  


**To Do** AF suggests splitting the continuous Rv load data into zero-low-high or similar categories? ref bimodal distributions in Fig 3b
