==> 04_model-analysis_status_infection-endpoint.Rmd <==
---
title: "04_model-analysis_infection-endpoint_status"
author: "Bryony Allen"
date: "09/05/2019"
output: word_document
---

```{r packages, include=FALSE, warning=FALSE, error=FALSE, message=FALSE}
# Install & load in required packages 

# for tidying data  
library(dplyr)
# library(tidyr)
# library(tibble)

# for plotting 
library(ggplot2) 
library(lattice)

# for stats 
library(binom)
library(multcomp)  ## WARNING this package masks dplyr::select
library(arm)

# library(purr)  # purr guidance > for evaluating models http://ijlyttle.github.io/isugg_purrr/presentation.html#(1) 
# library(broom)   # broom summarizes key information about models in tidy tibble()s >> https://github.com/tidymodels/broom
```

# Part 1: import data 

> **N.B.**   At the moment this code chuck reads in a .csv file with endpoint infection data, originally compiled in excel and then cleaned in "02_tidy_data-qpcr"script. In the future you will read in the .csv file of merged qPCR outputs (created, checked and cleaned in "02_tidy_data-qpcr"script) and experiment metadata (checked and cleaned in "02_tidy_data-metadata"script). 

<br>
```{r import data, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE}
data.endpoint <- read.csv('data/02_clean-data.csv') 

glimpse(data.endpoint)   #check how the dataset has imported (aka whether it has the data type right)  
```

# Part 2: visualise data 

```{r import proportion data , include=FALSE, warning=FALSE, error=FALSE, message=FALSE}
prop.infect.sum <- read.csv('data/03_prop-sum_data.csv') 
```

```{r graph labels, include=FALSE, warning=FALSE, error=FALSE, message=FALSE}
# make label vectors to use in plots
sp.labs <- c(Bb = "Bufo bufo", Rt = "Rana temporaria", Am = "Alytes muletensis")
exp.labs <- c("1" = "Bufo bufo I", "2" = "Bufo bufo II", "3" = "Rana temporaria", "4" = "Alytes muletensis","5" = "Alytes muletensis II")
exp.short.labs <- c("1" = "Bb1", "2" = "Bb2", "3" = "Rt", "4" = "Am")
infect.labs <- c(Bd = "Bd infection", Rv = "Rv infection")
expos.labs <- c("1" = "Bd", "2" = "Rv", "3" = "Bd-Rv", "4"="Rv-Bd")

```

```{r proportion infected by species, echo=FALSE, fig.cap= "Fig.1. Proportion of individuals infected, by pathogen, within a treatment group for the three host species. Please note that Alytes II has been excluded.", fig.width=16, fig.height=10}
plot.endpoint.prop.sp <- 
  prop.infect.sum %>% 
  filter(!ExperimentNo=='5')  %>% 
    ggplot(aes(x=Treatment, y= proportion.infected, fill=InfectType ))+ 
      geom_bar(position= "dodge", stat="identity") + 
      scale_x_discrete(name ="Treatment Group", limits=c("Bd","Rv","Bd-Rv", "Rv-Bd")) +       #change the order of the x axis ticks
      geom_errorbar(aes(ymin=upper, ymax=lower), width= .2, position= position_dodge(0.9)) +   #alpha to change transparancy 
      facet_grid(.~ExperimentNo, labeller=labeller(ExperimentNo = exp.labs)) + 
      theme(strip.text.x = element_text(size=12, face="italic")) +
      labs(x="Treatment Group",y="Proportion Infected") + 
      guides(fill=guide_legend(title="Pathogen")) 


plot.endpoint.prop.sp + theme(legend.position="top",legend.justification='right', legend.direction='horizontal') + theme(legend.key=element_blank()) + theme(axis.text.x = element_text(angle=30, hjust=1,vjust=1)) + theme(axis.title = element_text(size=14))

```

> **N.B.** *Bufo bufo* II have 2 individuals with Rv infection status in the Rv only group. These two records are sketchy as the qPCR results were inconclusive first time and then low the second time. Interestingly, the *Alytes muletensis* have the same pattern of a few individuals infected with ranavirus in the Rv only group with an equivalent number in the Bd-Rv treatment group.   ?? susceptibility pattern or timing of Rv dose pattern ???  


# Part 3: Bd: apply models to endpoint infection status  

Binomial GLM's where    
<br>
    response variable = Endpoint status [binary; 0,1]   
  <br>
    explanatory variable(s) = Treatment  [categorical]   &   ExperimentNo. [categorical]

<br>
  
  
**N.B.**  I use ExperimentNo as a proxy for species where

Experiment No. | Species              | total Bd zsp's   | min. temp. (^o^C) | max. temp. (^o^C)
---------- |------------------------- | -------------| -------------|------------- 
1 | *Bufo bufo* I          | 3,675,000    | 16.6        | 23.5     
2 | *Bufo bufo* II         | 1,443,750   | 16.7        | 27.6    
3 | *Rana temporaria*     | 2,336,250     | 16.7         | 27.6    
4 | *Alytes muletensis* I  | 472,500      | 15     | 16.6 
5 | *Alytes muletensis* II | 294,759      | 15     | 16.6 

... as this also accounts for Bd dose and room temperature variation between experiments. 


## Part 3a: Endpoint Infection Status: **Bd** 

Here I create a dataframe with the Rv-only treatment group removed (as they have never been exposed to Bd) and without *Alytes muletensis* babies as they only have one treatment group which I analyse seperately later.   

<br>
```{r Bd status df, results='hide', warning=FALSE, error=FALSE, message=FALSE}
Bd.status <- data.endpoint %>%
  filter(!Treatment=="Rv") %>%   # removal of Rv only treatment group 
  filter(!ExperimentNo=='5')  %>%   # removal of Alytes babies
  mutate(ExperimentNo = as.factor(ExperimentNo)) %>%
  dplyr::select(ID, Species, ExperimentNo, Treatment, Bd.endpoint.status, Bd.endpoint.GE) 

unique(data.endpoint$Treatment)
droplevels(Bd.status)
```

## Part 3d: Bd: model checks 

Here I check the model with the best fit
```bayesglm(Bd.endpoint.status ~ ExperimentNo, data=Bd.status, family=binomial)```

As we have separation, where a predictor perfectly predicts a binary response variable. The ```bayesglm``` function handles seperation 

> **N.B** there are no diagnostic plots for a Bayesian GLM

```{r Bd - bayes.glm - summary, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE}
 Bd.status.bayes <- bayesglm(Bd.endpoint.status ~ ExperimentNo, data=Bd.status, family=binomial)
summary(Bd.status.bayes)
```



## Part 3b: Bd: model fit 

```{r Bd fitted values, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE}
# To see the fitted values from a regression object (the values of the dependent variable predicted by the model), access the ```fitted.values``` attribute from a regression object with ````$fitted.values```.

names(Bd.status.bayes)   # look at the components of the glm object

Bd.status$bi.glm <- Bd.status.bayes$fitted.values  # add logisitic fitted values back to the dataframe as a new col

head(Bd.status) 
```

## Part 3c: Bd: model plotting   

> *Note* To plot the model you need a range of values for which to produce fitted values. Then use the ```predict()``` function to create the model for all the values. ```predict()``` gives you the predicted values based on your (fitted) linear model, the argument type="response" will give you the predicted probabilities 

```{r Bd - bi.glm - predicted values, include=TRUE, warning=FALSE, error=FALSE, message=FALSE}

# create a dataframe of "new" data 
newdat <- expand.grid(ExperimentNo=c("1", "2", "3", "4"),Treatment=c("Bd", "Bd-Rv", "Rv-Bd"))

# predict the value/result of the new data using the glm
newdat <-cbind(newdat, predict(object = Bd.status.bayes,   # the model 
                               newdata=newdat, se=TRUE, type="response", print.matrix=T))  # dataframe of new data 
newdat

expl.var <- c(1:3) # chose the range for the x-axis (Treatment)
exp.labs <- c("1" = "Bufo bufo I", "2" = "Bufo bufo II", "3" = "Rana temporaria", "4" = "Alytes muletensis")

# subset the data so you can plot each seperatly 
newdat1<- subset(newdat, ExperimentNo== "1")    
newdat2<- subset(newdat, ExperimentNo=="2")
newdat3<- subset(newdat, ExperimentNo=="3")
newdat4<- subset(newdat, ExperimentNo=="4")

```


```{r Bd - bi.glm - plot, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.cap= "Fig. 2. Probability of Bd infection status (GE) predicted by model"}
Bd.status.predict <- ggplot(newdat, aes(x= expl.var, y= fit, color=ExperimentNo)) +       # plot model estimates, color= the data you subsetted by
  geom_line(data = newdat1, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat1
  geom_errorbar(data = newdat1, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1)  +      # error bars for subset newdat1
  geom_line(data = newdat2, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat2
  geom_errorbar(data = newdat2, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1)  +      # error bars for subset newdat2
  geom_line(data = newdat3, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat3
  geom_errorbar(data = newdat3, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1)  +      # error bars for subset newdat3
  geom_line(data = newdat4, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat4
  geom_errorbar(data = newdat4, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1) +
  scale_x_continuous(breaks=seq(1:3),labels=c("Bd", "Bd-Rv", "Rv-Bd"))    # sets the breaks at 1,2 and 3 which correspond to the label names
      
Bd.status.predict.plot <- Bd.status.predict + 
    labs(title = "glm(Bd.endpoint.status ~ ExperimentNo,\n family = binomial)", x = "Treatment Group", y = "Bd status predictions\n(fit)", color = "Species\n") +
 scale_color_hue(labels = c("Bufo bufo I", "Bufo bufo II", "Rana temporaria", "Alytes muletensis I")) +
  theme_bw() +
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black")) 

Bd.status.predict.plot

# ggsave("figs/04_Bd_status-predict.png", plot= Bd.status.predict.plot, device=NULL) # export plot as .png 

```

**To Do** play with plots - check plot with Experiment No. as explanatory variable  


# Part 4: Bd: post-hoc tests 

Tukey's honest significant difference 
How to interpret the plots "Comparisons having intervals that do not overlap the vertical dashed line are significanty different. 
The vertical dashed line indicates no difference between the mean values for the factor-level comparisons indicated on the y-axis." ~ R BOOK  

```{r Bd - posthoc - , echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.width=8, fig.height=5}
# Bd.status4 <- glm(Bd.endpoint.status ~ ExperimentNo, data=Bd.status, family=binomial)

Tukey.Bd.status <- glht(Bd.status.bayes, linfct=mcp(ExperimentNo='Tukey'))
summary(Tukey.Bd.status) 
```
<br>
```{r Bd - posthoc - plot , echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.width=8, fig.height=5}
plot(Tukey.Bd.status)

# export plot as .png 
# png("figs/04_Bd_status-posthoc.png", type="cairo", units="in", width= 16, height=8, pointsize= 12, res=1000)
# par(mar=c(2,8,2,1))   #change the size of the margins the first number referes to the bottom margin, the second to the left margin, and so on in a clockwise fashion
# print(plot(Tukey.Bd.status))
# dev.off()
```

Where 
"1" = "Bufo bufo I",
"2" = "Bufo bufo II", 
"3" = "Rana temporaria", 
"4" = "Alytes muletensis"

**Conclusions**
The Tukey summary suggests everything is significant appart from the two *Bufo* experiments. 

The plots suggest: 

 * 2-1 : no significant difference between *Bufo bufo* experiments 
 * 3-2 & 3-1 : significant difference between *Rana* and both *Bufo bufo* experiments, with Rana less likely to be infected 
 * 4-2 & 4-1: no significant difference between *Alytes* and both *Bufo bufo* experiments 
 * 4-3: significant difference between *Alytes* and *Rana*, with Altyes more likely to be infected  


<br>


# Part 5: Rv: : apply models to endpoint infection status  

Here I create a dataframe with the Bd-only treatment group removed (as they have never been exposed to Rv) and without *Alytes muletensis* babies as they don't have Rv tratment groups.   

```{r Rv status df, results='hide', warning=FALSE, error=FALSE, message=FALSE}
Rv.status <- data.endpoint %>%
  filter(!Treatment=="Bd") %>%
  filter(!ExperimentNo=='5')  %>% 
  mutate(ExperimentNo = as.factor(ExperimentNo)) %>% 
  dplyr::select(ID, Species, ExperimentNo, Scenario, Treatment, Rv.MCPendpoint.status, Rv.endpoint.load) 
droplevels(Rv.status)

```

<br>

## Part 5a: Rv: model selection

The four models: 
```{r Rv glm}
Rv.status.glm <- glm(Rv.MCPendpoint.status ~ Treatment, data=Rv.status, family=binomial)

summary(Rv.status.glm)
```
```{r Rv - bi.glm - checks plot, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE}
par(mfrow=c(2,2), mar=c(3,3,3,1), mgp=c(2,0.8,0))
plot(Rv.status.glm)
```

## Part 5c: Rv: model fit

```{r Rv fitted values, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE}
# To see the fitted values from a regression object (the values of the dependent variable predicted by the model), access the ```fitted.values``` attribute from a regression object with ````$fitted.values```.

Rv.status$bi.glm <- Rv.status.glm$fitted.values  # add logisitic fitted values back to the dataframe as a new col
```

?? **QUESTION** It looks like the model is struggling to predict the probability of Rv infection status accurately ((**edit** *... in the Rv only treatment group. I agree that the predictions for both the coinfection groups match the raw data*)) 

## Part 5d: Rv: model plotting

```{r Rv - bi.glm - predictors, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE}

# create a dataframe of "new" data 
Rv.newdat <- expand.grid(ExperimentNo=c("1", "2", "3", "4"),Treatment=c("Bd-Rv", "Rv", "Rv-Bd"))

# predict the value/result of the new data using the glm
Rv.newdat <-cbind(Rv.newdat, predict(object = Rv.status.glm,   # the model 
                      newdata=Rv.newdat, se=TRUE, type="response", print.matrix=T))  # dataframe of new data 
Rv.newdat

expl.var <- c(1:3) # chose the range for the x-axis (Treatment)
exp.labs <- c("1" = "Bufo bufo I", "2" = "Bufo bufo II", "3" = "Rana temporaria", "4" = "Alytes muletensis")

newdat1<- subset(Rv.newdat, ExperimentNo== "1")    # need to subset the data so you can plot each seperatly 
newdat2<- subset(Rv.newdat, ExperimentNo=="2")
newdat3<- subset(Rv.newdat, ExperimentNo=="3")
newdat4<- subset(Rv.newdat, ExperimentNo=="4")
```

```{r Rv - bi.glm - plot, echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.cap= "Fig. 3. Probability of Rv infection status (GE) predicted by model"}
Rv.status.predict <- ggplot(Rv.newdat, aes(x= expl.var, y= fit)) +       # plot model estimates, color= the data you subsetted by
  geom_line(data = newdat1, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat1
  geom_errorbar(data = newdat1, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1)  +      # error bars for subset newdat1
  geom_line(data = newdat2, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat2
  geom_errorbar(data = newdat2, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1)  +      # error bars for subset newdat2
  geom_line(data = newdat3, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat1
  geom_errorbar(data = newdat3, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1)  +      # error bars for subset newdat1
  geom_line(data = newdat4, aes(x= expl.var, y= fit), size=1) +                                   # add lines for subset newdat1
  geom_errorbar(data = newdat4, aes(ymin=fit-se.fit, ymax=fit+se.fit), width=.03, size=1) +   scale_x_continuous(breaks=seq(1:3),labels=c("Bd-Rv", "Rv", "Rv-Bd"))    # sets the breaks at 1,2 and 3 which correspond to the label names

Rv.status.predict.plot <- Rv.status.predict + 
  ylab("Rv status predictions\n(fit)") +            # TO DO: would be good to have the axis marking 0 & 1 
  xlab("Treatment Group") +
  ggtitle("glm(Rv.MCPendpoint.status ~ Treatment, family = binomial)") +
  theme_bw() +
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black")) 

Rv.status.predict.plot

# ggsave("figs/04_Rv_status-predict.png", plot= Rv.status.predict.plot, device=NULL)  # export plot as .png 
```

<br>

# Part 6: Rv: post-hoc tests 

Tukey's honest significant difference 

```{r Rv - posthoc - , echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.width=8, fig.height=5}
# Rv.status3 <- glm(Rv.MCPendpoint.status ~ Treatment, data=Rv.status, family=binomial)

Tukey.Rv.status <- glht(Rv.status.glm, linfct=mcp(Treatment='Tukey'))
summary(Tukey.Rv.status) 
```
<br>
```{r Rv - posthoc - plot , echo=FALSE, warning=FALSE, error=FALSE, message=FALSE, fig.width=8, fig.height=5}

par(mar=c(2,8,2,1))   #change the size of the margins the first number referes to the bottom margin, the second to the left margin, and so on in a clockwise fashion
plot(Tukey.Rv.status)

# export plot as .png 
#png("figs/04_Rv_status-posthoc.png", type="cairo", units="in", width= 16, height=8, pointsize= 12, res=1000)
#par(mar=c(2,8,2,1))   #change the size of the margins the first number referes to the bottom margin, the second to the left margin, and so on in a clockwise fashion
#print(plot(Tukey.Rv.status))
#dev.off()
```

**Conclusion** 

- Rv vs. Bd-Rv is not significant 
- Rv-Bd vs. Bd-Rv there is significant difference between the coinfection groups 
- Rv-Bd vs. Rv is significant 

?? **QUESTION** is this highlighting the timing/sequence of dose again ?? 

<br>
